feat(swarm-client-behaviour): extract the client composite behaviour into its own crate
