+++ Cargo.toml
@@ -69,6 +69,7 @@ members = [
 
     # Node (libp2p boundary)
     "crates/swarm/client-protocol",
+    "crates/swarm/client-behaviour",
     "crates/swarm/node",
     "crates/swarm/topology",
     "crates/swarm/peers/peer-manager",
@@ -218,6 +219,7 @@ vertex-swarm-test-utils = { path = "crates/swarm/test-utils" }
 
 # node (libp2p boundary)
 vertex-swarm-client-protocol = { path = "crates/swarm/client-protocol" }
+vertex-swarm-client-behaviour = { path = "crates/swarm/client-behaviour" }
 vertex-swarm-node = { path = "crates/swarm/node" }
 
 # topology (in swarm/)
