+++ crates/swarm/node/src/protocol/forward.rs
@@ -1,30 +1,11 @@
-//! The forwarder seam: relay a retrieval or a pushsync to a closer peer.
+//! The concrete network forwarder: the real multi-hop relay for retrieval and
+//! pushsync.
 //!
-//! Inbound serving is handler-inline: each inbound retrieval or pushsync request
-//! becomes one self-contained future, with the substream itself as the
-//! correlation (mirroring the outbound model). When the local cache cannot
-//! answer a retrieval, or for every pushsync, the handler hands off to a
-//! [`Forwarder`] that relays to a closer peer and returns the result.
-//!
-//! Two implementations live here:
-//!
-//! - [`StubForwarder`] always returns [`ForwardError::NoCloserPeer`], so a cache
-//!   miss and every pushsync reset the inbound substream. It is the right
-//!   behaviour for a node that holds no reserve and takes no custody and never
-//!   wants to relay (and is the model the behaviour-level tests drive).
-//! - [`NetworkForwarder`] is the real multi-hop relay: it selects the closest
-//!   peer to the target excluding the requester (and ourselves), enforces the
-//!   forwarding-Kademlia loop rule (never relay to a peer that is not strictly
-//!   closer to the target, by XOR distance, than both the requester and this
-//!   node), reuses the existing self-contained outbound futures
-//!   ([`ClientHandle::retrieve_chunk`](crate::ClientHandle::retrieve_chunk) /
-//!   [`push_chunk`](crate::ClientHandle::push_chunk)) for the upstream leg, and
-//!   accounts both legs through the prepare/apply reservation actions so a
-//!   forwarder earns the spread. A failed forward drops both reservation actions
-//!   (release-on-drop), so no accounting leaks. Termination comes from the
-//!   strictly-closer rule (XOR distance decreases monotonically, bounded by the
-//!   address width), not from a hop/TTL counter, which the protocol does not
-//!   carry; [`MAX_FORWARD_CANDIDATES`] only caps per-node retry fan-out.
+//! The accounting-free forwarder seam (the [`Forwarder`] trait, [`StubForwarder`],
+//! the [`ForwardedChunk`] / [`ForwardedReceipt`] carriers, [`ForwardError`], and
+//! the [`closer_candidates`] selector) lives in `vertex-swarm-client-behaviour`.
+//! [`NetworkForwarder`] stays here because it couples to client accounting and
+//! the outbound [`ClientHandle`], which the behaviour crate must not depend on.
 //!
 //! Address verification lives at both edges. The downstream chunk returned by an
 //! upstream retrieval is verified against the requested address here (an
@@ -35,212 +16,32 @@
 //! signature-recovered owner for single-owner chunks; it does **not** validate
 //! the postage stamp's
 //! funding or expiry, which is a separate postage concern. A relayed pushsync
-//! receipt arrives as a [`Receipt`] whose storer was already recovered and
+//! receipt arrives as a [`Receipt`](vertex_swarm_net_pushsync::Receipt) whose
+//! storer was already recovered and
 //! verified at the decode boundary; the forwarder checks its custody depth
 //! before relaying, so `PO(storer, chunk)` must reach a depth derived from the
 //! locally observed neighbourhood depth. A forwarder never launders a shallow
 //! receipt upstream and scores the downstream peer adversely when it tries.
-//!
-//! # Upstream credit is deferred to the wire write
-//!
-//! The two legs are not both committed inside the forwarder. The downstream
-//! `receive` leg is genuinely complete the moment a verified chunk/receipt is in
-//! hand, so it is applied here. The upstream `provide` leg (the requester or
-//! pusher paying us) is *not* applied here: it is returned to the handler as a
-//! boxed [`AccountingAction`] and committed only after the chunk or receipt is
-//! successfully written back to the requester's substream. If that wire write
-//! fails, the handler drops the action, releasing the reservation, so the
-//! requester is never charged for a delivery it did not receive.
 
 use std::sync::Arc;
 
 use futures::future::BoxFuture;
-use nectar_primitives::{AnyChunk, ChunkAddress};
+use nectar_primitives::ChunkAddress;
 use tracing::{debug, warn};
 use vertex_swarm_api::{
     AccountingAction, PeerReporter, ReportSource, SwarmClientAccounting, SwarmScoringEvent,
     SwarmTopologyRouting, SwarmTopologyState,
 };
-use vertex_swarm_net_pushsync::{DepthVerdict, Receipt};
-use vertex_swarm_primitives::{OverlayAddress, Stamp, StampedChunk};
+use vertex_swarm_client_behaviour::{
+    ForwardError, ForwardedChunk, ForwardedReceipt, Forwarder, closer_candidates,
+};
+use vertex_swarm_net_pushsync::DepthVerdict;
+use vertex_swarm_primitives::{OverlayAddress, StampedChunk};
 
 use crate::ClientHandle;
 
 /// Report source for shallow/malformed receipts caught on the relay path.
 const PUSHSYNC_SOURCE: ReportSource = ReportSource::Protocol("pushsync");
-
-/// Maximum number of closer peers this node tries, in order, for a single
-/// inbound request before giving up.
-///
-/// This is a per-node fan-out cap, **not** a per-request hop/TTL counter: it
-/// bounds how many downstream candidates this one node retries for one inbound
-/// request, not the length of the overall A->B->C->... relay chain. Termination
-/// of the chain comes from the strictly-closer rule (every hop must hand the
-/// request to a peer strictly closer to the target by XOR distance than both the
-/// requester and this node), which makes proximity monotonically increase toward
-/// the target and is bounded by the address width, so no per-request hop counter
-/// or visited set is needed. The reference also walks a small fixed number of
-/// closer peers per hop; we mirror that with a bounded candidate set.
-const MAX_FORWARD_CANDIDATES: usize = 3;
-
-/// Why a forward could not complete.
-///
-/// The reason is intentionally coarse: the handler only needs to know the
-/// forward did not produce a chunk or receipt so it can reset the inbound
-/// substream. A real forwarder carries richer diagnostics for its own metrics,
-/// but the inbound serving path treats every failure as a reset.
-#[derive(Debug, thiserror::Error, strum::IntoStaticStr)]
-#[strum(serialize_all = "snake_case")]
-pub(crate) enum ForwardError {
-    /// No peer strictly closer to the target than both the requester and this
-    /// node is available to relay to. Covers both the no-candidate case and the
-    /// loop-prevention case (every candidate would forward sideways or backwards
-    /// in distance).
-    #[error("no closer peer to forward to")]
-    NoCloserPeer,
-
-    /// Every strictly-closer candidate we tried failed to answer.
-    #[error("all closer peers failed to relay")]
-    AllPeersFailed,
-
-    /// The upstream leg returned a chunk that does not answer the requested
-    /// address. Never relayed.
-    #[error("upstream relay returned unverified data")]
-    UnverifiedRelay,
-
-    /// The upstream leg returned a custody receipt whose recovered storer is not
-    /// deep enough for the chunk (`PO(storer, chunk) < required`). A forwarder
-    /// must never launder a shallow receipt upstream: it is dropped here and the
-    /// downstream peer is scored adversely. A malformed receipt does not reach
-    /// here; it is rejected at the downstream decode boundary and surfaces as a
-    /// push failure ([`AllPeersFailed`](Self::AllPeersFailed)).
-    #[error("upstream relay returned a shallow custody receipt")]
-    ShallowReceipt,
-
-    /// The upstream leg returned a custody receipt that cannot be judged because
-    /// the local neighbourhood view is not credible (the neighbourhood has not
-    /// saturated yet). The receipt is dropped without relaying it, but the
-    /// downstream peer is NOT penalised: the receipt may be honest, the local
-    /// node just lacks a trustworthy depth to anchor the check against. The relay
-    /// continues to the next candidate. Distinct from
-    /// [`ShallowReceipt`](Self::ShallowReceipt), which is a proven, penalised
-    /// finding of misbehaviour.
-    #[error("upstream relay returned an unverifiable custody receipt")]
-    UnverifiableReceipt,
-
-    /// Accounting refused one of the two legs (over the disconnect threshold),
-    /// so the relay was not attempted. Any reservation already taken is released
-    /// on drop.
-    #[error("accounting refused the relay")]
-    AccountingRefused,
-}
-
-/// Relays a retrieval or a pushsync to a closer peer on behalf of an inbound
-/// request.
-///
-/// `exclude` is the requester or pusher, passed so the forwarder never relays
-/// back to the peer that asked (loop prevention) and so it can account the
-/// inbound leg against it. The returned futures are `'static`, boxed, and `Send`
-/// so the handler can hold them in its inbound set: a libp2p `ConnectionHandler`
-/// is `Send` on both native and wasm (the browser `Stream` is itself `Send`), so
-/// the inbound serving futures are `Send` too.
-pub(crate) trait Forwarder: Send + Sync {
-    /// Retrieve `address` from a closer peer, excluding `exclude`.
-    ///
-    /// On success the downstream `receive` leg is already committed (we did
-    /// receive the chunk), and the un-applied upstream `provide` action is
-    /// returned alongside the chunk: the handler commits it only after the chunk
-    /// is written back to the requester, and drops it (releasing the
-    /// reservation) if that wire write fails.
-    fn retrieve(
-        &self,
-        address: ChunkAddress,
-        exclude: OverlayAddress,
-    ) -> BoxFuture<'static, Result<ForwardedChunk, ForwardError>>;
-
-    /// Push `chunk` to a closer peer, excluding `exclude`, returning the
-    /// storer's receipt to relay verbatim.
-    ///
-    /// The upstream `provide` action is returned un-applied for the same
-    /// deferred-commit reason as [`retrieve`](Self::retrieve).
-    fn push(
-        &self,
-        chunk: StampedChunk,
-        exclude: OverlayAddress,
-    ) -> BoxFuture<'static, Result<ForwardedReceipt, ForwardError>>;
-}
-
-/// A relayed chunk together with the un-applied upstream credit.
-///
-/// The forwarder hands this to the handler, which writes `chunk` back to the
-/// requester and only then commits `provide` (or drops it on a wire-write
-/// failure, releasing the reservation).
-///
-/// The stamp is optional: a storer answers a retrieval with the chunk bytes and
-/// may omit the stamp from the delivery. A stampless relayed chunk is served
-/// onward stampless (the requester validates it against its address, which is
-/// stamp-independent) and is not cached, since the cache value requires a stamp.
-pub(crate) struct ForwardedChunk {
-    /// The verified chunk to write back to the requester.
-    pub(crate) chunk: AnyChunk,
-    /// The stamp the downstream peer attached, if any.
-    pub(crate) stamp: Option<Stamp>,
-    /// The un-applied upstream credit; commit after a successful wire write.
-    pub(crate) provide: Box<dyn AccountingAction>,
-}
-
-impl std::fmt::Debug for ForwardedChunk {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("ForwardedChunk")
-            .field("chunk", &self.chunk)
-            .field("stamp", &self.stamp)
-            .finish_non_exhaustive()
-    }
-}
-
-/// A relayed receipt together with the un-applied upstream credit.
-pub(crate) struct ForwardedReceipt {
-    /// The storer's verified receipt to relay verbatim to the pusher. Its storer
-    /// was recovered at the downstream decode boundary, so a forwarder only
-    /// checks the depth policy before relaying.
-    pub(crate) receipt: Receipt,
-    /// The un-applied upstream credit; commit after a successful wire write.
-    pub(crate) provide: Box<dyn AccountingAction>,
-}
-
-impl std::fmt::Debug for ForwardedReceipt {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("ForwardedReceipt")
-            .field("receipt", &self.receipt)
-            .finish_non_exhaustive()
-    }
-}
-
-/// A forwarder that never relays: every relay fails with
-/// [`ForwardError::NoCloserPeer`].
-///
-/// A cache miss therefore resets the inbound retrieval substream and every
-/// inbound pushsync resets too, which is the correct behaviour for a node that
-/// holds no reserve, takes no custody, and does not participate as a relay.
-pub(crate) struct StubForwarder;
-
-impl Forwarder for StubForwarder {
-    fn retrieve(
-        &self,
-        _address: ChunkAddress,
-        _exclude: OverlayAddress,
-    ) -> BoxFuture<'static, Result<ForwardedChunk, ForwardError>> {
-        Box::pin(async { Err(ForwardError::NoCloserPeer) })
-    }
-
-    fn push(
-        &self,
-        _chunk: StampedChunk,
-        _exclude: OverlayAddress,
-    ) -> BoxFuture<'static, Result<ForwardedReceipt, ForwardError>> {
-        Box::pin(async { Err(ForwardError::NoCloserPeer) })
-    }
-}
 
 /// The real multi-hop relay: forwarding Kademlia for retrieval and pushsync.
 ///
@@ -254,14 +55,12 @@ impl Forwarder for StubForwarder {
 /// Forwarding Kademlia routes a request strictly toward the target's
 /// neighbourhood: each hop must hand the request to a peer that is *strictly
 /// closer* to the target than the peer that asked **and** strictly closer than
-/// this node itself, measured by full XOR distance (the same `closer` rule the
-/// reference uses, not capped proximity order). [`closer_candidates`] filters the
-/// topology's proximity-ordered candidates down to exactly those, also excluding
-/// the requester and ourselves. Because XOR distance to the target strictly
-/// decreases along the chain, a request can never cycle back to a peer it has
-/// already visited, so no per-request visited set or hop counter is needed; the
-/// chain is bounded by the address width. [`MAX_FORWARD_CANDIDATES`] only caps
-/// the per-node retry fan-out, not the chain length.
+/// this node itself, measured by full XOR distance. [`closer_candidates`] filters
+/// the topology's proximity-ordered candidates down to exactly those, also
+/// excluding the requester and ourselves. Because XOR distance to the target
+/// strictly decreases along the chain, a request can never cycle back to a peer
+/// it has already visited, so no per-request visited set or hop counter is
+/// needed; the chain is bounded by the address width.
 ///
 /// # Two-leg accounting
 ///
@@ -269,12 +68,11 @@ impl Forwarder for StubForwarder {
 /// we provide service to) and a *downstream* peer (the closer peer we relay to,
 /// the one that provides service to us). It credits the upstream leg
 /// (`prepare_provide_chunk(exclude)`) and debits the downstream leg
-/// (`prepare_receive_chunk(closer)`), exactly as the reference does, so it earns
-/// the price spread between the two. Both actions reserve on creation; on a
-/// successful relay both are applied, committing the balance changes. On any
-/// failure both actions are dropped, which releases the reservations
-/// (`ReceiveAction`/`ProvideAction` release on drop), so a failed forward never
-/// leaks an accounting reservation.
+/// (`prepare_receive_chunk(closer)`) so it earns the price spread between the
+/// two. Both actions reserve on creation; on a successful relay both are applied,
+/// committing the balance changes. On any failure both actions are dropped, which
+/// releases the reservations (`ReceiveAction`/`ProvideAction` release on drop),
+/// so a failed forward never leaks an accounting reservation.
 pub(crate) struct NetworkForwarder<T, A> {
     /// Our own overlay: excluded from candidates and used as the
     /// strictly-closer reference for the loop bound alongside the requester.
@@ -309,38 +107,6 @@ impl<T, A> NetworkForwarder<T, A> {
             reporter,
         }
     }
-}
-
-/// Select the peers strictly closer to `target` than both `requester` and
-/// `local`, excluding the requester and `local`, in proximity order, capped at
-/// [`MAX_FORWARD_CANDIDATES`].
-///
-/// This is the loop-prevention core, and it mirrors the reference's `closer`
-/// rule using **full XOR distance**, not capped proximity order. A candidate is
-/// kept only when it is strictly closer to the target than the requester (so the
-/// request never routes sideways or backwards and can never cycle) **and**
-/// strictly closer than this node (so a node already in the chunk's
-/// neighbourhood does not relay sideways to an equally deep peer, matching the
-/// reference's "closer than me" gate and avoiding the capped-PO plateau where all
-/// deep peers compare equal). Using XOR distance rather than capped proximity
-/// also lets the strict comparison distinguish peers inside the deepest band.
-fn closer_candidates(
-    topology: &impl SwarmTopologyRouting,
-    target: &ChunkAddress,
-    requester: OverlayAddress,
-    local: OverlayAddress,
-) -> Vec<OverlayAddress> {
-    topology
-        .closest_to(target, MAX_FORWARD_CANDIDATES * 2)
-        .into_iter()
-        .filter(|peer| *peer != requester && *peer != local)
-        // `target.closer(peer, other)` is true iff `peer` is strictly closer to
-        // `target` than `other` by full XOR distance. The candidate must beat
-        // both the requester (loop prevention) and this node (the reference's
-        // self-relative "closer than me" gate).
-        .filter(|peer| target.closer(peer, &requester) && target.closer(peer, &local))
-        .take(MAX_FORWARD_CANDIDATES)
-        .collect()
 }
 
 impl<T, A> Forwarder for NetworkForwarder<T, A>
@@ -388,11 +154,11 @@ where
 
                 match handle.retrieve_chunk(closer, address).await {
                     Ok(result) => {
-                        // Edge verification (#287): the relayed chunk must
-                        // answer the requested address before we account, cache,
-                        // or relay it. The chunk is address-derived (BMT hash or
-                        // owner plus signature), so equality proves it answers
-                        // the request, independent of the stamp. The handler
+                        // Edge verification: the relayed chunk must answer the
+                        // requested address before we account, cache, or relay
+                        // it. The chunk is address-derived (BMT hash or owner
+                        // plus signature), so equality proves it answers the
+                        // request, independent of the stamp. The handler
                         // re-checks the same equality before the wire.
                         if *result.chunk.address() == address {
                             // The downstream leg is genuinely complete (we
@@ -566,7 +332,7 @@ mod tests {
         SwarmScoringEvent,
     };
     use vertex_swarm_identity::Identity;
-    use vertex_swarm_net_pushsync::WireReceipt;
+    use vertex_swarm_net_pushsync::{Receipt, WireReceipt};
     use vertex_swarm_primitives::{Bin, StorageRadius};
     use vertex_swarm_spec::Spec;
     use vertex_swarm_test_utils::{MockTopology, test_identity_arc};
@@ -683,75 +449,6 @@ mod tests {
         ));
         let pricer = FixedPricer::new(10_000, vertex_swarm_spec::init_mainnet());
         Arc::new(ClientAccounting::new(bandwidth, pricer))
-    }
-
-    #[test]
-    fn closer_candidates_keeps_only_strictly_closer_peers() {
-        let address = *stamped().address();
-        // Requester shares 4 leading bits with the target.
-        let requester = overlay_at_proximity(&address, 4);
-        let closer = overlay_at_proximity(&address, 10);
-        let sideways = overlay_at_proximity(&address, 4);
-        let farther = overlay_at_proximity(&address, 1);
-        let local = OverlayAddress::from([0xee; 32]);
-
-        let topo = MockTopology::default().with_closest(vec![closer, sideways, farther]);
-        let got = closer_candidates(&topo, &address, requester, local);
-        assert_eq!(got, vec![closer], "only the strictly-closer peer survives");
-    }
-
-    #[test]
-    fn closer_candidates_excludes_requester_and_local() {
-        let address = *stamped().address();
-        let requester = overlay_at_proximity(&address, 4);
-        // Local is farther from the target than the candidate, so the
-        // self-relative gate does not reject the candidate; this isolates the
-        // exclusion behaviour (local must be dropped because it is us, not
-        // because of the closer-than-me gate).
-        let local = overlay_at_proximity(&address, 2);
-        let closer = overlay_at_proximity(&address, 10);
-
-        // The topology returns local and requester among the closest; both must
-        // be filtered out as self/requester.
-        let topo = MockTopology::default().with_closest(vec![local, closer, requester]);
-        let got = closer_candidates(&topo, &address, requester, local);
-        assert_eq!(got, vec![closer]);
-    }
-
-    #[test]
-    fn closer_candidates_drops_peers_farther_than_local() {
-        // The reference's self-relative gate: a node already deeper in the
-        // target's neighbourhood than a candidate must not relay sideways/back to
-        // that candidate, even when the candidate is still closer than the
-        // requester. This is the divergence-from-reference fix.
-        let address = *stamped().address();
-        let requester = overlay_at_proximity(&address, 2);
-        // We (local) share 12 bits with the target.
-        let local = overlay_at_proximity(&address, 12);
-        // The candidate is closer than the requester (8 > 2) but farther than us
-        // (8 < 12), so it must be rejected.
-        let farther_than_local = overlay_at_proximity(&address, 8);
-        // A candidate deeper than us survives.
-        let deeper = overlay_at_proximity(&address, 20);
-
-        let topo = MockTopology::default().with_closest(vec![deeper, farther_than_local]);
-        let got = closer_candidates(&topo, &address, requester, local);
-        assert_eq!(
-            got,
-            vec![deeper],
-            "only a peer strictly closer than this node survives"
-        );
-    }
-
-    #[test]
-    fn closer_candidates_empty_when_nothing_is_closer() {
-        let address = *stamped().address();
-        let requester = overlay_at_proximity(&address, 12);
-        let local = OverlayAddress::from([0xee; 32]);
-        let sideways = overlay_at_proximity(&address, 8);
-
-        let topo = MockTopology::default().with_closest(vec![sideways]);
-        assert!(closer_candidates(&topo, &address, requester, local).is_empty());
     }
 
     /// Drive a forwarder future to completion while answering the single
@@ -1063,7 +760,7 @@ mod tests {
         assert!(matches!(err, ForwardError::ShallowReceipt));
 
         // The downstream peer that handed us the shallow receipt is scored as
-        // invalid data through the same reporter #287 uses.
+        // invalid data through the same reporter the inbound edge uses.
         let (reported_peer, event, source) = reporter.single();
         assert_eq!(reported_peer, closer);
         assert_eq!(event, SwarmScoringEvent::InvalidData);
@@ -1136,11 +833,11 @@ mod tests {
 
     #[tokio::test]
     async fn push_with_non_credible_view_is_unverifiable_and_does_not_penalise() {
-        // Regression for #316: with a non-credible local view (the neighbourhood
-        // has not saturated) the forwarder cannot judge custody depth, so even a
-        // shallow receipt declaring radius 0 must not be relayed AND the
-        // downstream peer must not be penalised. The forward fails with
-        // `UnverifiableReceipt`, nothing is scored, and no reservation leaks.
+        // With a non-credible local view (the neighbourhood has not saturated)
+        // the forwarder cannot judge custody depth, so even a shallow receipt
+        // declaring radius 0 must not be relayed AND the downstream peer must not
+        // be penalised. The forward fails with `UnverifiableReceipt`, nothing is
+        // scored, and no reservation leaks.
         let chunk = stamped();
         let address = *chunk.address();
         let pusher = overlay_at_proximity(&address, 2);
